a function that creat a file

+++ 0x15-file_io/1-create_file.c
@@ -8,9 +8,9 @@
  * Return: c
 */
 
-int _strlen(char *s)
+ssize_t _strlen(char *s)
 {
-	int c = 0;
+	ssize_t c = 0;
 
 	while (*s != '\0')
 	{
